refactor(orchestrator): extract namespace read fanout into orchestration/namespace-read-fanout.ts (seam 26, #1526)
Behavior-preserving move of the namespace-scoped READ fanout (401 LOC,
11 methods, 17 deps): storage-dir hint loading, cross-namespace search
and memory/archive reads, scoped candidate search, artifact recall
fanout, artifact source-status resolution (with its WeakMap cache), and
path/storage-dir namespace mapping helpers. Namespace RESOLUTION stays
with the #1521 ScopePlan resolver — this coordinator only fans reads
out. ARTIFACT_STATUS_CACHE_TTL_MS static widened from private.

orchestrator.ts: 4,864 -> 4,594 LOC.
