+++ packages/remnic-core/src/orchestrator.ts
@@ -116,6 +116,7 @@ import { RecallIntrospectionCoordinator } from "./orchestration/recall-introspec
 import { OrchestratorInitCoordinator } from "./orchestration/orchestrator-init.js";
 import { PersistenceIndexCoordinator } from "./orchestration/persistence-index.js";
 import { WorkspaceOpsCoordinator } from "./orchestration/workspace-ops.js";
+import { NamespaceReadFanoutCoordinator } from "./orchestration/namespace-read-fanout.js";
 import {
   abortRecallError,
   buildCompressionGuidelinesMarkdown,
@@ -682,7 +683,8 @@ export class Orchestrator {
       statuses: Map<string, "active" | "superseded" | "archived" | "missing">;
     }
   >();
-  private static readonly ARTIFACT_STATUS_CACHE_TTL_MS = 60_000;
+  /** Read by NamespaceReadFanoutCoordinator (seam 26), hence not `private`. */
+  static readonly ARTIFACT_STATUS_CACHE_TTL_MS = 60_000;
 
   // Access tracking buffer (Phase 1A)
   // Maps memoryId -> {count, lastAccessed} for batched updates
@@ -1249,98 +1251,17 @@ export class Orchestrator {
     resolvedStorageDir: string,
     configured: Set<string>,
   ): { namespace: string; identityToken: string; storageDir: string } {
-    const currentRank = this.namespaceStorageDirHintOwnershipRank(
+    return this.namespaceReadFanoutCoordinator.preferNamespaceStorageDirHintOwner(
       current,
-      resolvedStorageDir,
-      configured,
-    );
-    const candidateRank = this.namespaceStorageDirHintOwnershipRank(
       candidate,
       resolvedStorageDir,
       configured,
     );
-    if (candidateRank < currentRank) return candidate;
-    if (candidateRank > currentRank) return current;
-
-    const byName = candidate.namespace.localeCompare(current.namespace);
-    if (byName < 0) return candidate;
-    if (byName > 0) return current;
-    return candidate.identityToken.localeCompare(current.identityToken) < 0
-      ? candidate
-      : current;
   }
 
   private loadNamespaceStorageDirHintsFromCatalog(): void {
-    if (this.namespaceStorageDirHintsLoaded || !this.namespaceCatalog.enabled) return;
-    this.namespaceStorageDirHintsLoaded = true;
-    const catalogPath = path.join(this.config.memoryDir, "state", "namespaces.jsonl");
-    if (!existsSync(catalogPath)) return;
-
-    let body: string;
-    try {
-      body = readFileSync(catalogPath, "utf8");
-    } catch {
-      return;
-    }
-
-    const compactedByNamespace = new Map<
-      string,
-      { namespace: string; identityToken: string; storageDir: string }
-    >();
-    for (const line of body.split(/\r?\n/)) {
-      const trimmed = line.trim();
-      if (!trimmed) continue;
-      try {
-        const parsed = JSON.parse(trimmed) as unknown;
-        if (!parsed || typeof parsed !== "object" || Array.isArray(parsed)) continue;
-        const record = parsed as Record<string, unknown>;
-        if (
-          typeof record.namespace !== "string" ||
-          typeof record.storageDir !== "string" ||
-          typeof record.identityToken !== "string"
-        ) {
-          continue;
-        }
-        const namespace = normalizeNamespaceIdentity(record.namespace);
-        if (!namespace || record.identityToken !== namespaceIdentityToken(namespace)) continue;
-        compactedByNamespace.set(namespace, {
-          namespace,
-          identityToken: record.identityToken,
-          storageDir: record.storageDir,
-        });
-      } catch {
-        // Catalog hints are best-effort. The catalog reader still owns full recovery.
-      }
-    }
-
-    const configured = new Set(
-      this.configuredNamespaceList().map((namespace) => normalizeNamespaceIdentity(namespace)),
+    return this.namespaceReadFanoutCoordinator.loadNamespaceStorageDirHintsFromCatalog(
     );
-    const preferredByStorageDir = new Map<
-      string,
-      { namespace: string; identityToken: string; storageDir: string }
-    >();
-    for (const record of compactedByNamespace.values()) {
-      if (!this.storageDirMatchesNamespaceHint(record.namespace, record.storageDir)) {
-        continue;
-      }
-      const resolvedStorageDir = path.resolve(record.storageDir);
-      const current = preferredByStorageDir.get(resolvedStorageDir);
-      preferredByStorageDir.set(
-        resolvedStorageDir,
-        current
-          ? this.preferNamespaceStorageDirHintOwner(
-              current,
-              record,
-              resolvedStorageDir,
-              configured,
-            )
-          : record,
-      );
-    }
-    for (const record of preferredByStorageDir.values()) {
-      this.rememberNamespaceStorageDirHint(record.namespace, record.storageDir);
-    }
   }
 
   private async maintenanceNamespaces(
@@ -1416,68 +1337,9 @@ export class Orchestrator {
     searchOptions?: SearchQueryOptions;
     execution?: SearchExecutionOptions;
   }): Promise<QmdSearchResult[]> {
-    if (
-      resolveNamespaceCapabilities(this.config).namespaces &&
-      options.namespaces !== undefined &&
-      options.namespaces.length === 0
-    ) {
-      return [];
-    }
-    const namespaces = resolveNamespaceCapabilities(this.config).namespaces
-      ? Array.from(
-          new Set(
-            (options.namespaces?.length
-              ? options.namespaces
-              : this.configuredNamespaceList()
-            )
-              .map((value) => value.trim())
-              .filter(Boolean),
-          ),
-        )
-      : [this.config.defaultNamespace];
-
-    if (!resolveNamespaceCapabilities(this.config).namespaces) {
-      switch (options.mode) {
-        case "hybrid":
-          return await this.qmd.hybridSearch(
-            options.query,
-            undefined,
-            options.maxResults,
-            options.execution,
-          );
-        case "bm25":
-          return await this.qmd.bm25Search(
-            options.query,
-            undefined,
-            options.maxResults,
-            options.execution,
-          );
-        case "vector":
-          return await this.qmd.vectorSearch(
-            options.query,
-            undefined,
-            options.maxResults,
-            options.execution,
-          );
-        default:
-          return await this.qmd.search(
-            options.query,
-            undefined,
-            options.maxResults,
-            options.searchOptions,
-            options.execution,
-          );
-      }
-    }
-
-    return await this.namespaceSearchRouter.searchAcrossNamespaces({
-      query: options.query,
-      namespaces,
-      maxResults: options.maxResults,
-      mode: options.mode,
-      searchOptions: options.searchOptions,
-      execution: options.execution,
-    });
+    return this.namespaceReadFanoutCoordinator.searchAcrossNamespaces(
+      options,
+    );
   }
 
   async searchHealthForNamespace(
@@ -2024,92 +1886,10 @@ export class Orchestrator {
     storage: StorageManager,
     sourceIds: string[],
   ): Promise<Map<string, "active" | "superseded" | "archived" | "missing">> {
-    const currentStatusVersion = storage.getMemoryStatusVersion();
-    const cached = this.artifactSourceStatusCache.get(storage);
-    let snapshot = cached;
-    const isFresh =
-      snapshot !== undefined &&
-      Date.now() - snapshot.loadedAtMs <=
-        Orchestrator.ARTIFACT_STATUS_CACHE_TTL_MS &&
-      snapshot.statusVersion === currentStatusVersion;
-
-    const rebuildSnapshot = async () => {
-      const MAX_STABLE_READ_ATTEMPTS = 3;
-      let latestStatuses = new Map<
-        string,
-        "active" | "superseded" | "archived" | "missing"
-      >();
-      let latestVersionAfter = storage.getMemoryStatusVersion();
-
-      for (let attempt = 0; attempt < MAX_STABLE_READ_ATTEMPTS; attempt += 1) {
-        const versionBefore = storage.getMemoryStatusVersion();
-        const allMemories = await storage.readAllMemories();
-        const versionAfter = storage.getMemoryStatusVersion();
-        latestVersionAfter = versionAfter;
-        latestStatuses = new Map(
-          allMemories.map((m) => [
-            m.frontmatter.id,
-            (m.frontmatter.status ?? "active") as
-              | "active"
-              | "superseded"
-              | "archived"
-              | "missing",
-          ]),
-        );
-
-        if (versionAfter === versionBefore) {
-          const rebuilt = {
-            loadedAtMs: Date.now(),
-            statusVersion: versionAfter,
-            statuses: latestStatuses,
-          };
-          this.artifactSourceStatusCache.set(storage, rebuilt);
-          return rebuilt;
-        }
-      }
-
-      // Sustained write churn: return latest read without caching a potentially torn snapshot.
-      return {
-        loadedAtMs: Date.now(),
-        statusVersion: latestVersionAfter,
-        statuses: latestStatuses,
-      };
-    };
-
-    if (!isFresh) {
-      snapshot = await rebuildSnapshot();
-    } else {
-      // Warm cache may miss brand-new sourceMemoryId values created after snapshot build.
-      // Refresh once on-demand when unseen IDs are requested.
-      const hasUnknownSourceIds = sourceIds.some(
-        (id) => !snapshot?.statuses.has(id),
-      );
-      if (hasUnknownSourceIds) {
-        snapshot = await rebuildSnapshot();
-      }
-    }
-
-    // Persist negative lookups in the cached snapshot so stale source IDs do not
-    // trigger repeated full snapshot rebuilds on every matching recall.
-    for (const id of sourceIds) {
-      if (!snapshot?.statuses.has(id)) {
-        snapshot?.statuses.set(id, "missing");
-      }
-    }
-
-    const statuses = new Map<
-      string,
-      "active" | "superseded" | "archived" | "missing"
-    >();
-    for (const id of sourceIds) {
-      const status = snapshot?.statuses.get(id);
-      if (status) {
-        statuses.set(id, status);
-      } else {
-        statuses.set(id, "missing");
-      }
-    }
-    return statuses;
+    return this.namespaceReadFanoutCoordinator.resolveArtifactSourceStatuses(
+      storage,
+      sourceIds,
+    );
   }
 
   /**
@@ -2931,15 +2711,11 @@ export class Orchestrator {
     recallNamespaces: string[],
     targetCount: number,
   ): Promise<MemoryFile[]> {
-    if (targetCount <= 0) return [];
-    const namespaces = Array.from(new Set(recallNamespaces));
-    const filteredByNamespace = await Promise.all(
-      namespaces.map((namespace) =>
-        this.fetchActiveArtifactsForNamespace(namespace, prompt, targetCount),
-      ),
+    return this.namespaceReadFanoutCoordinator.recallArtifactsAcrossNamespaces(
+      prompt,
+      recallNamespaces,
+      targetCount,
     );
-
-    return mergeArtifactRecallCandidates(filteredByNamespace, targetCount);
   }
 
   private scopeQueryAwarePaths(
@@ -2979,51 +2755,12 @@ export class Orchestrator {
       allowArchived?: boolean;
     },
   ): Promise<QmdSearchResult[]> {
-    const cappedLimit = Math.max(0, limit);
-    if (cappedLimit === 0 || candidatePaths.size === 0) return [];
-
-    const tokens = Array.from(new Set(tokenizeRecallQuery(query)));
-    const memories = (
-      await Promise.all(
-        Array.from(candidatePaths).map(async (memoryPath) => {
-          const namespace = resolveNamespaceCapabilities(this.config).namespaces
-            ? this.namespaceFromPath(memoryPath)
-            : this.config.defaultNamespace;
-          const storage = await this.storageRouter.storageFor(namespace);
-          return await storage.readMemoryByPath(memoryPath);
-        }),
-      )
-    ).filter((memory): memory is MemoryFile => memory !== null);
-
-    const results: QmdSearchResult[] = [];
-    for (const memory of memories) {
-      const status = memory.frontmatter.status ?? "active";
-      if (!options?.allowArchived && status !== "active") continue;
-
-      const haystack = [
-        memory.content,
-        memory.frontmatter.category,
-        ...(memory.frontmatter.tags ?? []),
-      ]
-        .join(" ")
-        .toLowerCase();
-      let hits = 0;
-      for (const token of tokens) {
-        if (haystack.includes(token)) hits += 1;
-      }
-      const score = tokens.length > 0 ? hits / tokens.length : 0.01;
-      if (tokens.length > 0 && hits === 0) continue;
-
-      results.push({
-        docid: memory.frontmatter.id,
-        path: memory.path,
-        score,
-        snippet: memory.content.slice(0, 400).replace(/\n/g, " "),
-        transport: "scoped_prefilter",
-      });
-    }
-
-    return results.sort((a, b) => b.score - a.score).slice(0, cappedLimit);
+    return this.namespaceReadFanoutCoordinator.searchScopedMemoryCandidates(
+      candidatePaths,
+      query,
+      limit,
+      options,
+    );
   }
 
   private async fetchQmdMemoryResultsWithArtifactTopUp(
@@ -3117,26 +2854,9 @@ export class Orchestrator {
   private async resolveStateDirForNamespace(
     namespace: string,
   ): Promise<string> {
-    if (!resolveNamespaceCapabilities(this.config).namespaces) {
-      return path.join(this.config.memoryDir, "state");
-    }
-    if (namespace !== this.config.defaultNamespace) {
-      return path.join(this.config.memoryDir, "namespaces", namespace, "state");
-    }
-    const candidate = path.join(
-      this.config.memoryDir,
-      "namespaces",
-      this.config.defaultNamespace,
+    return this.namespaceReadFanoutCoordinator.resolveStateDirForNamespace(
+      namespace,
     );
-    try {
-      const candidateStat = await stat(candidate);
-      if (candidateStat.isDirectory()) {
-        return path.join(candidate, "state");
-      }
-    } catch {
-      // Fall back to the legacy root when the migrated default namespace directory is absent.
-    }
-    return path.join(this.config.memoryDir, "state");
   }
 
   private buildGraphRecallRankedResults(
@@ -3582,6 +3302,41 @@ export class Orchestrator {
       });
     }
     return this._workspaceOpsCoordinator;
+  }
+
+  /**
+   * Namespace read-fanout coordinator (issue #1526 seam 26). Owns
+   * namespace-scoped read fanout (hints, cross-namespace search/reads,
+   * artifact statuses). Lazy + accessor-wired (late-binding rule).
+   */
+  private _namespaceReadFanoutCoordinator: NamespaceReadFanoutCoordinator | undefined;
+
+  private get namespaceReadFanoutCoordinator(): NamespaceReadFanoutCoordinator {
+    if (!this._namespaceReadFanoutCoordinator) {
+      // eslint-disable-next-line @typescript-eslint/no-this-alias
+      const self = this;
+      this._namespaceReadFanoutCoordinator = new NamespaceReadFanoutCoordinator({
+        get artifactSourceStatusCache() { return self.artifactSourceStatusCache; },
+        get config() { return self.config; },
+        configuredNamespaceList: () => self.configuredNamespaceList(),
+        fetchActiveArtifactsForNamespace: (namespace, prompt, targetCount) => self.fetchActiveArtifactsForNamespace(namespace, prompt, targetCount),
+        loadNamespaceStorageDirHintsFromCatalog: () => self.loadNamespaceStorageDirHintsFromCatalog(),
+        get namespaceCatalog() { return self.namespaceCatalog; },
+        namespaceFromPath: (p) => self.namespaceFromPath(p),
+        get namespaceSearchRouter() { return self.namespaceSearchRouter; },
+        namespaceStorageDirHintOwnershipRank: (record, resolvedStorageDir, configured) => self.namespaceStorageDirHintOwnershipRank(record, resolvedStorageDir, configured),
+        get namespaceStorageDirHints() { return self.namespaceStorageDirHints; },
+        get namespaceStorageDirHintsLoaded() { return self.namespaceStorageDirHintsLoaded; },
+        set namespaceStorageDirHintsLoaded(value) { self.namespaceStorageDirHintsLoaded = value; },
+        preferNamespaceStorageDirHintOwner: (current, candidate, resolvedStorageDir, configured) => self.preferNamespaceStorageDirHintOwner(current, candidate, resolvedStorageDir, configured),
+        get qmd() { return self.qmd; },
+        qmdCollectionNamespaceFromPrefix: (collectionPrefix) => self.qmdCollectionNamespaceFromPrefix(collectionPrefix),
+        rememberNamespaceStorageDirHint: (namespace, storageDir) => self.rememberNamespaceStorageDirHint(namespace, storageDir),
+        storageDirMatchesNamespaceHint: (namespace, storageDir) => self.storageDirMatchesNamespaceHint(namespace, storageDir),
+        get storageRouter() { return self.storageRouter; },
+      });
+    }
+    return this._namespaceReadFanoutCoordinator;
   }
 
   private async recallInternal(
@@ -4790,29 +4545,15 @@ export class Orchestrator {
   }
 
   private namespaceFromPath(p: string): string {
-    if (!resolveNamespaceCapabilities(this.config).namespaces) return this.config.defaultNamespace;
-    const parts = qmdCollectionPathParts(p);
-    const collectionNamespace = parts
-      ? this.qmdCollectionNamespaceFromPrefix(parts.collection)
-      : null;
-    if (collectionNamespace) return collectionNamespace;
-    const m = p.match(/[\\/]+namespaces[\\/]+([^\\/]+)(?:[\\/]|$)/);
-    if (!m?.[1]) return this.config.defaultNamespace;
-    return namespaceIdentityFromToken(m[1]) ?? m[1];
+    return this.namespaceReadFanoutCoordinator.namespaceFromPath(
+      p,
+    );
   }
 
   private storageDirNamespace(storageDir: string): string {
-    // #1521: delegates to the scope-module resolver. The inline dir→namespace
-    // derivation (token round-trip guard, catalog hints) is retired so the
-    // adHocNamespaceResolutions ratchet no longer counts this site. Hints are
-    // loaded lazily via the callback (only after early returns, matching the
-    // original behavior — codex P2).
-    return resolveNamespaceFromStorageDir(storageDir, {
-      config: this.config,
-      configuredNamespaces: this.configuredNamespaceList(),
-      hints: this.namespaceStorageDirHints,
-      loadHints: () => this.loadNamespaceStorageDirHintsFromCatalog(),
-    });
+    return this.namespaceReadFanoutCoordinator.storageDirNamespace(
+      storageDir,
+    );
   }
 
   // #1522: catalog touch methods removed — touches now happen at the storage chokepoint.
@@ -4838,26 +4579,16 @@ export class Orchestrator {
   private async readAllMemoriesForNamespaces(
     namespaces: string[],
   ): Promise<MemoryFile[]> {
-    const uniq = Array.from(new Set(namespaces.filter(Boolean)));
-    const lists = await Promise.all(
-      uniq.map(async (ns) => {
-        const sm = await this.storageRouter.storageFor(ns);
-        return sm.readAllMemories();
-      }),
+    return this.namespaceReadFanoutCoordinator.readAllMemoriesForNamespaces(
+      namespaces,
     );
-    return lists.flat();
   }
 
   private async readArchivedMemoriesForNamespaces(
     namespaces: string[],
   ): Promise<MemoryFile[]> {
-    const uniq = Array.from(new Set(namespaces.filter(Boolean)));
-    const lists = await Promise.all(
-      uniq.map(async (ns) => {
-        const sm = await this.storageRouter.storageFor(ns);
-        return sm.readArchivedMemories();
-      }),
+    return this.namespaceReadFanoutCoordinator.readArchivedMemoriesForNamespaces(
+      namespaces,
     );
-    return lists.flat();
   }
 }
